Use Carthage to import CareKit and ResearchKit
Currently, CareKit is not on CocoaPods. Additionally, there
is a critical bugfix sitting in a PR- so I'm pulling this
from my local drive. Finally, ResearchKit has a bug that
prevents it from working with CocoaPods 1.0.0. That PR
has been merged but waiting for Apple to push the fix to
CocoaPods. When all these issues are resolved, we should
be able to rely only on CocoaPods, but for now Carthage
is needed.